backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
openai>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Utilities
requests>=2.31.0
